Compute the frequency entropy using scikit-maad

Compute different entropies based on the average spectrum, its variance, and its maxima.

Usage

maad_spectral_entropy(object, flim = NULL, maad = NULL)

Arguments

object

A Wave object or a spectrogram_maad object generated by maad_spectrogram. If a Wave-like object is provided, the spectrogram will be calculated using the default parameters.

flim

A numeric vector of length 2 specifying the frequency limits for the entropy calculation.

maad

An optional maad object. If not provided, one will be created using getMaad().

Value

List comprising:

EAS

Entropy of Average Spectrum.

ECU

Entropy of spectral variance (along the time axis for each frequency).

ECV

Entropy of Coefficient of Variation (along the time axis for each frequency).

EPS

Entropy of spectral maxima (peaks).

EPS_KURT

Kurtosis of spectral maxima.

EPS_SKEW

Skewness of spectral maxima.

Details

For addition documentation see https://scikit-maad.github.io/generated/maad.features.spectral_entropy.html