Dolbear's law

dolbear1897;textualsonicscrewdriver was the first to publish a formula for how the rate of chirping of crickets varies with temperature, using the tree cricket Oecanthus fultoni.

Usage

dolbear(n = NULL, t = NULL, species = "Oecanthus fultoni")

Arguments

n

Chirps per minute

t

Temperature in Celsius

species

Species to use (by default Oecanthus fultoni), NULL to calculate for all species.

Value

Data frame with t and n calculated for matching species.

Details

Subsequent research by Dolbear and others have published additional formulae for other species: Oecanthus fultoni fulton1925sonicscrewdriver, Oecanthus rileyi walker1962sonicscrewdriver, Oecanthus alexanderi walker2010sonicscrewdriver, and Oecanthus allardi allard1957sonicscrewdriver.

References

Examples

dolbear(n=6)
#>             species location      m      c min max       source        t n
#> 1 Oecanthus fultoni     Iowa 7.7879 -30.21  15  25 Fulton, 1925 4.649520 6
#> 2 Oecanthus fultoni   Oregon 9.2007 -36.53   9  28 Fulton, 1925 4.622474 6
#> 3 Oecanthus fultoni     Ohio 8.2080 -38.61  18  31 Walker, 1962 5.434942 6
dolbear(t=25)
#>             species location      m      c min max       source  t        n
#> 1 Oecanthus fultoni     Iowa 7.7879 -30.21  15  25 Fulton, 1925 25 164.4875
#> 2 Oecanthus fultoni   Oregon 9.2007 -36.53   9  28 Fulton, 1925 25 193.4875
#> 3 Oecanthus fultoni     Ohio 8.2080 -38.61  18  31 Walker, 1962 25 166.5900